:bug: read redis info from config

+++ modules/diary/config.go
@@ -4,5 +4,5 @@ type Config struct {
 	EnabledGroups []int64 `yaml:"enabled_groups"`
 	RedisAddr     string  `yaml:"redis_addr"`
 	RedisPassword string  `yaml:"redis_password"`
-	RedisDb       int64   `yaml:"redis_db"`
+	RedisDb       int     `yaml:"redis_db"`
 }
